Refactoring packages as mentioned by mentor
