+++ src/main/java/com/yourcodereview/jegors/task1/converter/IPv4Converter.java
@@ -1,0 +1,9 @@
+package com.yourcodereview.jegors.task1.converter;
+
+import java.util.function.ToIntFunction;
+
+/**
+ * Convert string representation of an IPv4 address to int.
+ */
+public interface IPv4Converter extends ToIntFunction<CharSequence> {
+}
