--- src/main/java/lv/id/jc/ipcounter/collector/package-info.java
@@ -1,4 +1,0 @@
-/**
- * The package contains the IP address collector.
- */
-package lv.id.jc.ipcounter.collector;
