--- src/main/java/lv/id/jc/ipcounter/converter/IPv4Converter.java
@@ -1,9 +1,0 @@
-package lv.id.jc.ipcounter.converter;
-
-import java.util.function.ToIntFunction;
-
-/**
- * Convert string representation of an IPv4 address to int.
- */
-public interface IPv4Converter extends ToIntFunction<CharSequence> {
-}
