--- src/main/java/lv/id/jc/ipcounter/converter/package-info.java
@@ -1,11 +1,0 @@
-/**
- * The package contains a converter for IPv4 addresses.
- * <p>
- * There is one simple implementations of the interface in the package:
- * <p>
- * {@link lv.id.jc.ipcounter.converter.IPv4SimpleConverter}
- * This converter converts the representation of an IPv4 address from a string type
- * to a number. The converter does not check the string representation of the ip address.
- * In the case of an erroneous address format, the returned number is unpredictable.
- */
-package lv.id.jc.ipcounter.converter;
